+++ tests/Helper/Ordered.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Neat\Object\Test\Helper;
+
+class Ordered
+{
+    const ORDER_BY = 'name';
+
+    /** @var int */
+    public $id;
+
+    /** @var string */
+    public $name;
+}